requirements.txt
ffmpeg
noisereduce
Pillow
pandas
google-cloud-speech
google-cloud-storage
#google